rebranding reactormonk -> scalanlp
