+++ opennlp/src/main/scala/org/scalanlp/epic-opennlp/NameFinder.scala
@@ -1,4 +1,4 @@
-package org.reactormonk.epic.opennlp
+package org.scalanlp.epic.opennlp
 
 import epic.slab._
 import epic.slab.Utils._
